fix(cron): resolve linter errors by deriving countdown state and isolating cron utilities
